+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+-   Rewritten test runner for better stability
+-   Now single steps are marked as not runnable
+-   Better error reporting when a step fails
+
 ## [0.3.0]
 
 -   Added support for gherkin context actions